chore(statistics): normalize year/month casts to ::smallint
The existence-guard and INSERT already use ::smallint, but the read-back
query was using ::int. This change makes all three sites consistent.

The cast change is functionally neutral (smallint widens to int
automatically in JS via Number()).

+++ apps/backend/src/modules/statistics/statistics.repository.ts
@@ -339,8 +339,8 @@ export class StatisticsRepository {
     `);
 
     const targetRows = await executor.execute<{ year: number; month: number }>(sql`
-      SELECT EXTRACT(YEAR FROM (now() - INTERVAL '1 month'))::int AS year,
-             EXTRACT(MONTH FROM (now() - INTERVAL '1 month'))::int AS month
+      SELECT EXTRACT(YEAR FROM (now() - INTERVAL '1 month'))::smallint AS year,
+             EXTRACT(MONTH FROM (now() - INTERVAL '1 month'))::smallint AS month
     `);
     const target = targetRows[0];
     if (!target) throw new Error('captureMonthlyIfNewMonth: target-month query returned no rows');
